Tree method in default.ini and other functions
Clear_execution function: Clear all the temporary files according to the executed flow
TreeMethod: the type of phylogenetic tree algorithm
