+++ config/default.ini
@@ -5,6 +5,7 @@ ScriptDir 		= %(localScript)s/scripts
 Environ			= config/parsl.env
 Workload		= config/work.config
 NetworkMethod   = MP
+TreeMethod      = IQTREE
 
 [WORKFLOW]
 Name			= BioWorkFlow
